+++ src/domains/overview/tableModel/tables/calculated/CalcSequence.tsx
@@ -1,8 +1,8 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { DavitIcons } from "../../../../../components/atomic/IconSet";
 import { DavitTableRowData } from "../../../../../components/common/fragments/DavitTable";
-import { DavitIcons } from "../../../../../components/common/IconSet";
 import { SequenceCTO } from "../../../../../dataAccess/access/cto/SequenceCTO";
 import { Terminal } from "../../../../../dataAccess/access/types/GoToType";
 import { CalculatedStep } from "../../../../../services/SequenceService";
